win lane LLP64 (b): host thunk signatures — i64 slots are int64_t, never long
Layer (b) of run-burndown defect #2: the extern-C runtime thunks the JIT
binds must agree with the CIR's respelled long-long externs, and host
`long` is 32-bit on a win64 build. Widened: madarray_size (returned
long), madarray_assign_int/_bool, __madc_sys_init/_once,
__madc_atomic_fetch_add_l (the CIR's wide arm), the _chk family size
params + object_size (size_t — the glibc contract), the packed-varargs
bridge's %d-family cast (the slot is 8 bytes), and cir_builder's own
storage-word model (sizeof/alignof(long) -> long long: the emitted
buffer element IS long long now; alignof(long)=4 on win64 under-aligned
8-byte slots).

Classified OUT (platform-long semantics, dialect question (c), task
#45): the *l bit-scan family (__madc_ffsl/clzl/...) and the generic
__madc_{add,sub,mul}_overflow triple — gcc's *l builtins operate on
TARGET long (32-bit on win64), so those follow the script-long width
decision, not the i64 law. The _s64/_u64 width-specific overflow
helpers are already long long. __madc_fd_set/timeval family: no
consumer anywhere in-tree (dead thunks, noted for cruft removal).
__madc_try_context_size returns unsigned long but has no in-tree
declarer; left for (c).

Hypothesis: thunks written as `long` on LP64 silently narrow to 32-bit
on a win64 host build while the CIR now passes 8-byte args.
Layer: JIT call (long long per CIR) -> import-resolved host thunk
(C++ signature) — the signature IS the deepest layer on the host side.
Searched: "extern-C thunk signatures with bare long" — grep extern-C
defs across madc_mir_backend/va_helpers/ns_*/parser/rt/ + the CIR's
i64-extern symbol list; __php_array_get_int, rt_except, madc_value_*,
and the libmadc eval API were already int64_t.
Oracle: n/a — LP64 no-op by construction (long == int64_t there); the
suite is the oracle; win64 validation = lld16/stdio15 reducers after
the exe rebuild.

+++ src/cir_builder.cpp
@@ -5182,7 +5182,7 @@ node_t CirBuilder::flavor_marshal_thunk_def(const char *thunk_sym,
 		snprintf(tnm, sizeof(tnm), "__fmt%zu", i);
 		tnames[i] = tnm;
 		stmts.push_back(obj_storage_decl(tnm, 4, dtor_sym.c_str(),
-						 NULL, alignof(long)));
+						 NULL, alignof(long long)));
 		node_t ca = list();
 		append(ca, temp_addr(tnm));
 		append(ca, member_call1(cstr_sym, anames[i].c_str()));
@@ -5279,7 +5279,7 @@ size_t CirBuilder::object_class_words(DataDefCLASS *cdd) const
 	if (cdd->size == 0) return 0;
 	if (cdd->is_complete && cdd->size == 1 && !cdd->has_vptr_slot)
 		return 0;
-	return (cdd->size + sizeof(long) - 1) / sizeof(long);
+	return (cdd->size + sizeof(long long) - 1) / sizeof(long long);
 }
 
 // N_TYPE node for a (void*) cast: TYPE(LIST(VOID), DECL(IGNORE, LIST(POINTER))).
@@ -5527,7 +5527,7 @@ node_t CirBuilder::obj_storage_decl(const char *name, size_t words,
 	// An object whose alignment exceeds the long[] buffer's natural 8
 	// (e.g. the 16-aligned madc_value inside madc::value) declares it:
 	// _Alignas(align) long name[words].
-	if (align > alignof(long))
+	if (align > alignof(long long))
 		append(spec, node1(N_ALIGNAS, integer((int64_t)align, origin)));
 	append_i64(spec, origin);
 	node_t share = node1(N_SHARE, spec);
@@ -5582,7 +5582,7 @@ bool CirBuilder::is_array_object(DataDef *dd)
 
 size_t CirBuilder::array_obj_words() const
 {
-	return (sizeof(madc::value) + sizeof(long) - 1) / sizeof(long);
+	return (sizeof(madc::value) + sizeof(long long) - 1) / sizeof(long long);
 }
 
 node_t CirBuilder::array_storage_decl(const char *name, TokenBase *origin)
@@ -9072,7 +9072,7 @@ node_t CirBuilder::member_node(const memberpair_t &m, DataDefSTRUCT *owner)
 	// User dims (if any) stay outermost, as in the class-object arm above.
 	if (is_array_object(mtype)) {
 		node_t mspec = list();
-		if (alignof(madc::value) > alignof(long))
+		if (alignof(madc::value) > alignof(long long))
 			append(mspec, node1(N_ALIGNAS,
 				integer((int64_t)alignof(madc::value), m.origin)));
 		append_i64(mspec, m.origin);
@@ -20307,7 +20307,7 @@ node_t CirBuilder::translate_expr(TokenBase *tb)
 					m_pending_stmts.push_back(
 						obj_storage_decl(tnm, 4,
 							hdtor.c_str(), tb,
-							alignof(long)));
+							alignof(long long)));
 					auto taddr = [&]() {
 						return node1(N_ADDR,
 							node2(N_IND, id(tnm, tb),
@@ -21265,8 +21265,8 @@ node_t CirBuilder::translate_try(TokenTRY *tt)
 
 	// Opaque, 8-aligned storage sized to MadcTryContext (jmp_buf + 2 ptrs). The
 	// builder is C++ so it knows the real ABI size; round up to a long[] count.
-	size_t ctx_words = (sizeof(jmp_buf) + 2 * sizeof(void *) + sizeof(long) - 1)
-			   / sizeof(long);
+	size_t ctx_words = (sizeof(jmp_buf) + 2 * sizeof(void *) + sizeof(long long) - 1)
+			   / sizeof(long long);
 	node_t ctx_spec = list();
 	append_i64(ctx_spec, tt);
 	node_t ctx_decl_list = list();
